# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RedeStore API",
    "description": "API da RedeStore: autenticação, loja (produtos e pedidos) e eventos (com inscrições). Endpoints protegidos exigem o header `Authorization: Bearer <token>`, obtido em `/auth/login` ou `/auth/cadastro`. Erros seguem o formato ProblemDetails (RFC 7807): `title` traz o código do erro (ex.: `EVENTO_NAO_ENCONTRADO`) e `detail` a mensagem.",
    "version": "v1"
  },
  "paths": {
...
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Retorna um usuário pelo id",
        "description": "O próprio usuário pode consultar o seu id; admin pode consultar qualquer um. Demais casos retornam 403 ACESSO_NEGADO.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/eventos": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Lista eventos",
        "description": "apenasFuturos=true retorna só eventos com dataHora >= agora (UTC). Padrão: false (todos).",
        "parameters": [
          {
            "name": "apenasFuturos",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Eventos"
        ],
        "summary": "Cria um evento (admin)",
        "description": "Regras: titulo, descricao, local e foto obrigatórios; preco >= 0; vagasTotais >= 1.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarEventoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/eventos/{id}": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Retorna um evento pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Eventos"
        ],
        "summary": "Atualiza parcialmente um evento (admin)",
        "description": "Só os campos enviados (não nulos) são alterados. Reduzir vagasTotais abaixo do número de inscrições confirmadas retorna 409 EVENTO_VAGAS_TOTAIS_INSUFICIENTES.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarEventoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Eventos"
        ],
        "summary": "Remove um evento (admin)",
        "description": "Só é possível remover eventos sem inscrições confirmadas; caso contrário retorna 409 EVENTO_COM_INSCRICOES_CONFIRMADAS.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/eventos/{id}/vagas-restantes": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Retorna quantas vagas restam no evento",
        "description": "vagasRestantes = vagasTotais - inscrições com status 'confirmada'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VagasRestantesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AtualizarEventoRequest": {
        "required": [
          "titulo",
          "descricao",
          "dataHora",
          "local",
          "preco",
          "vagasTotais",
          "foto"
        ],
        "type": "object",
        "properties": {
          "titulo": {
            "type": [
              "null",
              "string"
            ]
          },
          "descricao": {
            "type": [
              "null",
              "string"
            ]
          },
          "dataHora": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "local": {
            "type": [
              "null",
              "string"
            ]
          },
          "preco": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "vagasTotais": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "foto": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CriarEventoRequest": {
        "required": [
          "titulo",
          "descricao",
          "dataHora",
          "local",
          "preco",
          "vagasTotais",
          "foto"
        ],
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "dataHora": {
            "type": "string",
            "format": "date-time"
          },
          "local": {
            "type": "string"
          },
          "preco": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "vagasTotais": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "foto": {
            "type": "string"
          }
        }
      },
...
      "EventoDto": {
        "required": [
          "id",
          "titulo",
          "descricao",
          "dataHora",
          "local",
          "preco",
          "vagasTotais",
          "vagasRestantes",
          "foto"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "dataHora": {
            "type": "string",
            "format": "date-time"
          },
          "local": {
            "type": "string"
          },
          "preco": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "vagasTotais": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "vagasRestantes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "foto": {
            "type": "string"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UsuarioDto": {
        "required": [
          "id",
          "nome",
          "email",
          "telefone",
          "papel"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": [
              "null",
              "string"
            ]
          },
          "papel": {
            "type": "string"
          }
        }
      },
      "VagasRestantesDto": {
        "required": [
          "vagasRestantes"
        ],
        "type": "object",
        "properties": {
          "vagasRestantes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT retornado por /auth/login ou /auth/cadastro. Validade padrão: 8 horas.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}